# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"GDSC Resume AI Chat","version":"v1"},"paths":{"/common_question":{"get":{"tags":["questions"],"summary":"Get Common Questions","operationId":"get_common_questions_common_question_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/asked_question":{"get":{"tags":["answers"],"summary":"Asked Common Question","operationId":"asked_common_question_asked_question_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Asked Common Question Asked Question Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/submit_tech_answer/{question_id}":{"post":{"tags":["answers"],"summary":"Submit Tech Answer","operationId":"submit_tech_answer_submit_tech_answer__question_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"integer","title":"Question Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleAnswer"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submit_behav_answer/{question_id}":{"post":{"tags":["answers"],"summary":"Submit Behav Answer","operationId":"submit_behav_answer_submit_behav_answer__question_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"integer","title":"Question Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleAnswer"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submit_personal_answer":{"post":{"tags":["answers"],"summary":"Submit Personal Answer","operationId":"submit_personal_answer_submit_personal_answer_post","parameters":[{"name":"question_id","in":"query","required":true,"schema":{"type":"integer","title":"Question Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleAnswer"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SingleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/":{"post":{"tags":["auth"],"summary":"Create User","operationId":"create_user_auth__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cv":{"get":{"tags":["user"],"summary":"Get User Cv","operationId":"get_user_cv_cv_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["user"],"summary":"Upload Cv","operationId":"upload_cv_cv_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CVRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/cv/{cv_id}":{"put":{"tags":["user"],"summary":"Update Cv","operationId":"update_cv_cv__cv_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"cv_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Cv Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CVRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["user"],"summary":"Delete Cv","operationId":"delete_cv_cv__cv_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"cv_id","in":"path","required":true,"schema":{"type":"integer","exclusiveMinimum":0,"title":"Cv Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_for_access_token_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_auth_token_post"},"CVRequest":{"properties":{"content":{"type":"string","title":"Content"},"position":{"$ref":"#/components/schemas/Position"}},"type":"object","required":["content","position"],"title":"CVRequest"},"CreateUserRequest":{"properties":{"username":{"type":"string","title":"Username"},"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","name","email","password"],"title":"CreateUserRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Position":{"type":"string","enum":["ai","be","fe","mobile"],"title":"Position"},"QuestionType":{"type":"string","enum":["tech_q","behav_q","personal_q"],"title":"QuestionType"},"SingleAnswer":{"properties":{"answer":{"type":"string","title":"Answer","default":"I DON'T KNOW"}},"type":"object","title":"SingleAnswer"},"SingleResponse":{"properties":{"type":{"$ref":"#/components/schemas/QuestionType"},"question":{"type":"string","title":"Question"},"user_answer":{"type":"string","title":"User Answer"},"criteria":{"type":"string","title":"Criteria"},"evaluation":{"type":"string","title":"Evaluation"}},"type":"object","required":["type","question","user_answer"],"title":"SingleResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}